[FIX] varios sobre approval state y devoluciones
Part-of: ingadhoc/odoo-public-administration#567
